anda ldr_c.c

+++ solucion/src/ldr_c.c
@@ -19,7 +19,7 @@ void ldr_c    (
     unsigned char (*src_matrix)[src_row_size] = (unsigned char (*)[src_row_size]) src;
     unsigned char (*dst_matrix)[dst_row_size] = (unsigned char (*)[dst_row_size]) dst;
 
-    float max = 4876875;
+    double max = 4876875;
 
     for (int i = 0; i < filas; i++)
     {
